requirements/common.txt
Django<2.0
djangorestframework
drf-nested-routers
coreapi
stravalib
python-memcached
pymongo>=3.2
jsonpickle
celery>=3.1.23,<4.0
django-celery
